add WithStack
In some cases when errors can't be passed along a chain using
With/Wrap and friends it can be helpful to grab the stack and use
that. Such as when recovering from a panic.

To help that, add WithStack which works like With but builds an error
stack from the calling stack instead of only from the immediate
caller.